Wait only for the minimum number of peers
Before submitting genesis tx. For this purpose set A is formed out of the first minimum peers that are online.

Also some restructuring for genesis code.

+++ iroha_client/tests/multisignature_transaction.rs
@@ -14,7 +14,7 @@ mod tests {
     const CONFIGURATION_PATH: &str = "tests/test_config.json";
     const CLIENT_CONFIGURATION_PATH: &str = "tests/test_client_config.json";
     const N_PEERS: usize = 4;
-    const MAX_FAULTS: usize = 1;
+    const MAX_FAULTS: u32 = 1;
 
     #[test]
     //TODO: use cucumber_rust to write `gherkin` instead of code.
